requirements.txt
requests
requests-oauthlib
jsonobject
kivy
pirc522
pyyaml

